Node process finished
Leitura e execução dos nodes da folha salva no banco de dados no processo principal através dos botões de "start/stop | pause/resume"
